refactor: rename launches api to remote data source
Refactor naming to better align the android platform guides.
