+++ app/src/test/java/com/washuTechnologies/merced/util/MockRepositoryHelper.kt
@@ -1,7 +1,6 @@
 package com.washuTechnologies.merced.util
 
-import com.washuTechnologies.merced.api.launches.LaunchesApi
-import com.washuTechnologies.merced.api.launches.RocketLaunch
+import com.washuTechnologies.merced.api.launches.LaunchesRemoteDatasource
 import com.washuTechnologies.merced.api.launches.RocketLaunchRepository
 import com.washuTechnologies.merced.database.launches.LaunchesLocalDatasource
 import com.washuTechnologies.merced.datasources.ConnectivityDatasource
@@ -9,11 +8,11 @@ import com.washuTechnologies.merced.datasources.ConnectivityDatasource
 object RepositoryHelper {
 
     fun launchesRepository(
-        launchesRemoteDatasource: LaunchesApi = MockDatasourceHelper.launchesRemoteDatasource(),
+        launchesRemoteDatasource: LaunchesRemoteDatasource = MockDatasourceHelper.launchesRemoteDatasource(),
         localDatasource: LaunchesLocalDatasource = MockDatasourceHelper.mockLaunchesLocalSource(),
         connectivityDatasource: ConnectivityDatasource = MockDatasourceHelper.mockConnectivity()
     ) = RocketLaunchRepository(
-        launchesApi = launchesRemoteDatasource,
+        remoteDatasource = launchesRemoteDatasource,
         localDatasource = localDatasource,
         connectivityDatasource = connectivityDatasource
     )
